requirements.txt
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
requests==2.25.1
requests-file==1.5.1
soupsieve==2.3.1
sqlparse==0.5.3
psycopg2-binary==2.9.10